# BTTH/BTTH2/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BTTH2, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/BrandModels1": {
      "get": {
        "tags": [
          "BrandModels1"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BrandModels"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BrandModels1"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BrandModels"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BrandModels"
            }
          }
        }
      }
    },
    "/api/BrandModels1/{id}": {
      "get": {
        "tags": [
          "BrandModels1"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BrandModels"
            }
          }
        }
      },
      "put": {
        "tags": [
          "BrandModels1"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BrandModels"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BrandModels1"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CategoryModels1": {
      "get": {
        "tags": [
          "CategoryModels1"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CategoryModels"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CategoryModels1"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CategoryModels"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CategoryModels"
            }
          }
        }
      }
    },
    "/api/CategoryModels1/{id}": {
      "get": {
        "tags": [
          "CategoryModels1"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CategoryModels"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CategoryModels1"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CategoryModels"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "CategoryModels1"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ProductModels1": {
      "get": {
        "tags": [
          "ProductModels1"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductModels"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProductModels1"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductModels"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProductModels"
            }
          }
        }
      }
    },
    "/api/ProductModels1/{id}": {
      "get": {
        "tags": [
          "ProductModels1"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProductModels"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProductModels1"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductModels"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ProductModels1"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "BrandModels": {
      "required": [
        "brandDesc",
        "brandName",
        "brandOrder"
      ],
      "type": "object",
      "properties": {
        "brandId": {
          "format": "int32",
          "type": "integer"
        },
        "brandName": {
          "minLength": 1,
          "type": "string"
        },
        "brandDesc": {
          "minLength": 1,
          "type": "string"
        },
        "brandOrder": {
          "format": "int32",
          "type": "integer"
        },
        "brand": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductModels"
          }
        }
      },
      "additionalProperties": false
    },
    "CategoryModels": {
      "required": [
        "cateDesc",
        "cateName",
        "cateOrder"
      ],
      "type": "object",
      "properties": {
        "cateId": {
          "format": "int32",
          "type": "integer"
        },
        "cateName": {
          "minLength": 1,
          "type": "string"
        },
        "cateDesc": {
          "minLength": 1,
          "type": "string"
        },
        "cateOrder": {
          "format": "int32",
          "type": "integer"
        },
        "category": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductModels"
          }
        }
      },
      "additionalProperties": false
    },
    "DateOnly": {
      "type": "object",
      "properties": {
        "year": {
          "format": "int32",
          "type": "integer"
        },
        "month": {
          "format": "int32",
          "type": "integer"
        },
        "day": {
          "format": "int32",
          "type": "integer"
        },
        "dayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "dayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      },
      "additionalProperties": false
    },
    "DayOfWeek": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "ProductModels": {
      "required": [
        "brandId",
        "cateId",
        "productDescription",
        "productImg",
        "productName",
        "productPrice",
        "productQuantity"
      ],
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "minLength": 1,
          "type": "string"
        },
        "productDescription": {
          "minLength": 1,
          "type": "string"
        },
        "productPrice": {
          "format": "double",
          "type": "number"
        },
        "productQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "productImg": {
          "minLength": 1,
          "type": "string"
        },
        "cateId": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 1,
          "type": "integer"
        },
        "category": {
          "$ref": "#/definitions/CategoryModels"
        },
        "brandId": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 1,
          "type": "integer"
        },
        "brand": {
          "$ref": "#/definitions/BrandModels"
        }
      },
      "additionalProperties": false
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "$ref": "#/definitions/DateOnly"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}